### 2021-04-30 #3223 @Kevin-Mattheus-Moerman
@isidentical thanks for this submission. I have a couple of questions about this submission which I hope you can help with.
[JOSS requires](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) software to have an obvious research application. Currently, from reading the paper and the statement of need, I get the impression this work presents a very valuable and useful tool for source code searching, however, I fail to understand if their is a clear and direct application to scientific research. Can you help clarify this? Perhaps you can update the statement of need section in the paper to include a clearer description of the scientific research applications. In essence the description should straight away convince a reader that one would definitively cite this work when using it for research because it is clearly important to research or the research outcomes.

If you make changes to the paper you can update the pdf here by calling `@whedon generate pdf`.
