### 2021-05-01 3d46ce0
utils: implement tqdm-less progress bars for docker
