+++ paper/source/evaluation.md
@@ -5,6 +5,8 @@ For the limited subset of things that can be described in any of the competitor 
 evaluate the performance of `Reiz` by running similiar queries in Github Code Search \[@githubcodesearch\]
 grep.app \[@grepapp\] and Krugle \[@krugle\] and report back the amount of true / false positives.
 
+## Comparisons
+
 The method of evaluation is running the queries and analyzing the top 10 results. Some engines offer
 multiple matches per result, and they will be marked as true positive if any of them are a true
 positive. We also discard match spans, since none of the competitors can successfully display the expression
@@ -24,6 +26,7 @@ Objective: search for an addition or a subtraction operation
 | engine               | query                          | true positives | false positives |
 |----------------------|--------------------------------|----------------|-----------------|
 | Github Code Search   | `language:python + -`          | 0              | 0               |
+| Krugle (fuzzy)       | `expr + expr` / `expr - expr`  | 0              | 0               |
 | Krugle (solr syntax) | `\+ \-`                        | 2              | 8               |
 | Krugle (regex syntax)| `(.*)(\+\|\-)(.*)`             | 1              | 9               |
 | grep.app             | `(.*)(\+\|\-)(.*)`             | 2              | 8               |
