### 2021-05-01 #3223 @isidentical
> Currently, from reading the paper and the statement of need, I get the impression this work presents a very valuable and useful tool for source code searching, however, I fail to understand if their is a clear and direct application to scientific research. Can you help clarify this?

Hey @Kevin-Mattheus-Moerman! Thanks for the response and the question. In this research, we share the findings and a practical implementation of a source code search engine that actually process the information regarding syntactical structs in a scalable way. I believe that the major scientific value comes from the new DSL that we introduce, and the query compiler for it.

I've updated the second parapraph of the state of the need describing it's scientific value.
