### 2021-05-08 4c0e0d5
docker: update edgedb to beta2
