### 2021-05-08 c55e563
workflow: dont run benchmarks on PRs
