+++ .github/workflows/test.yml
@@ -42,7 +42,7 @@ jobs:
         EDGEDB_SERVER_BIN: edgedb-server
       run: |
         coverage run tests/runner.py --change-db-schema \
-        --start-edgedb-server --run-benchmarks
+        --start-edgedb-server
         coverage xml
 
     - name: Upload coverage to Codecov
