### 2021-05-08 5eb5391
tests: output benchmark results in a separate file
