### 2021-05-19 #3223 @mjsottile
:wave: @lutzhamel, @yuhc.  Would you be interested in reviewing this submission to JOSS?  If you haven’t done a JOSS review before, we perform the reviews on github and the process is much lighter weight than a traditional paper review.  The review criteria that we ask reviewers to use as a guide are here (a helpful checklist is provided on the review issue for reference once the review starts): https://joss.readthedocs.io/en/latest/review_criteria.html

Let me know if you’re interested.  Thanks!
