### 2021-05-19 #3223 @lutzhamel
Hi Matthew:

I'd be more than happy to review this submission. Two quick questions:
what is the typical time frame for a JOSS review?  Does the length of the
submitted paper matter? In other words, is there a minimum or maximum
number of pages that should be considered?

Thanks,
Lutz

On Wed, May 19, 2021 at 4:53 AM Matthew Sottile ***@***.***>
wrote:

> 👋 @lutzhamel <https://github.com/lutzhamel>, @yuhc
> <https://github.com/yuhc>. Would you be interested in reviewing this
> submission to JOSS? If you haven’t done a JOSS review before, we perform
> the reviews on github and the process is much lighter weight than a
> traditional paper review. The review criteria that we ask reviewers to use
> as a guide are here (a helpful checklist is provided on the review issue
> for reference once the review starts):
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> Let me know if you’re interested. Thanks!
>
