### 2021-05-19 #3223 @mjsottile
Excellent - thanks for the positive responses.  In response to your questions:

- Timeframe is generally shorter than a classical journal paper, usually on the order of 1-2 months.  We used to go faster, but in the last year things have slowed down.
- The papers are usually short, anywhere from 3-6 pages usually.  Longer sometimes happens, but it’s generally discouraged unless absolutely necessary.  The review checklist has specific things to look for in the write up.
- The review organization is all based on github issue threads.  When I initiate the review later today, you’ll see a new issue will be created for the review itself.  At the top of the issue each reviewer will have a checklist to work through.  If a submission satisfied all things in the checklist and there are no issues for the author to address, a reviewer can just check off the items and post a note stating that they’re happy with the submission.  If there are issues, usually reviewers post them in the review thread and/or open issues on the project itself (e.g., if there are missing files, issues with the build, bugs noticed, etc…).  Typically we ask that reviewers just make note of the project issue in a comment on the review thread so we’re aware of them.  We don’t expect reviewers to write up a review like you’d see for a normal paper - the checklist and comments in the review thread are sufficient.

If you want to see some example reviews, just pick a paper that was published from the list (https://joss.theoj.org/papers/published) and you can see the papers themselves for a sense of scale/scope, and see the review threads that correspond to each paper.  Some are super short if the paper and package have no issues from the start, while others can get longer if there are more things too address before it’s publishable.
