### 2021-05-24 #3296 @yuhc
Hi @mjsottile, I've tried to run the software within the latest docker on Ubuntu 20.04, but it requires a subtle change and doesn't return search results as expected. Shall I contact the author to resolve this issue? (The document doesn't state enough about the environment settings.)
