### 2021-05-24 #3296 @yuhc
> Please let me know about any problems you experience to build / run the Reiz on the bug tracker @yuhc. Thanks!

Hi @isidentical, I'm running Ubuntu 20.04, Docker 19.03.13, docker-compose 1.26.2.
I had to change [docker-compose's version](https://github.com/reizio/reiz.io/blob/master/docker-compose.yml#L1) from 3.9 to 3.8 to build and start the container.

The build instruction is usually expected to contain the system environment.

I didn't see `API is running on ...` on the server end, but I could access the search engine from 8080 port (remotely, as I don't have a desktop environment) and tried to search `Call(Name("len"))`. However, the loading icon never disappeared, and no results was ever returned.

Could you explain more about what the dataset is built on top of ("~75 files from 10 different projects") and help check what's going wrong here?
Also, it'll be great to know the use cases of reiz.io.
