### 2021-05-24 #3296 @isidentical
> Hi @isidentical, I'm running Ubuntu 20.04, Docker 19.03.13, docker-compose 1.26.2.
I had to change docker-compose's version from 3.9 to 3.8 to build and start the container.

Interesting, I can successfully spin the instances with 3.9 though will definitely investigate (would you mind opening an issue on the tracker).

> I didn't see API is running on ... on the server end, but I could access the search engine from 8080 port (remotely, as I don't have a desktop environment) and tried to search Call(Name("len"))

You should wait the API to start before accessing, since without it the web ui will just wait and timeout eventually. Would you mind sending me the logs (btw it would be better if you could create an issue on the repo itself!)

Thanks!
