### 2021-05-24 reizio/reiz.io#51 @yuhc
Docker-compose version not supported
I'm running Ubuntu 20.04, Docker 19.03.13, docker-compose 1.26.2.
I had to change docker-compose's version from 3.9 to 3.8 to build and start the container.

The [docker compose compatibility matrix](https://docs.docker.com/compose/compose-file/) also doesn't contain a "3.9" entry.
