### 2021-05-24 reizio/reiz.io#52 @yuhc
Search engine hanging forever
The engine is started in the following manner:

```
ubuntu@ip-172-31-63-88:~/reiz.io$ docker-compose up --build --remove-orphans
Building reiz
Step 1/8 : FROM python:3.8.2-slim
 ---> e7d894e42148
Step 2/8 : RUN apt-get update  && apt-get upgrade -y  && apt-get install git apt-utils bash -y
 ---> Using cache
 ---> 882941d1de25
Step 3/8 : WORKDIR /app
 ---> Using cache
 ---> ae7e60c190f3
Step 4/8 : COPY requirements.txt requirements.txt
 ---> Using cache
 ---> 6a970b289647
Step 5/8 : RUN python -m pip install -r requirements.txt
 ---> Using cache
 ---> 395c80b3ae8f
Step 6/8 : COPY . .
 ---> b3d4d8bdf508
Step 7/8 : RUN python -m pip install -e .
 ---> Running in 91727f3854c6
Obtaining file:///app
  Installing build dependencies: started
  Installing build dependencies: finished with status 'done'
  Getting requirements to build wheel: started
  Getting requirements to build wheel: finished with status 'done'
    Preparing wheel metadata: started
    Preparing wheel metadata: finished with status 'done'
Installing collected packages: reiz
  Running setup.py develop for reiz
Successfully installed reiz
WARNING: You are using pip version 20.1; however, version 21.1.1 is available.
You should consider upgrading via the '/usr/local/bin/python -m pip install --upgrade pip' command.
Removing intermediate container 91727f3854c6
 ---> 6135f850cf1a
Step 8/8 : ENTRYPOINT ["/bin/bash", "./scripts/docker_bootstrap.sh"]
 ---> Running in 75c35cd4e076
Removing intermediate container 75c35cd4e076
 ---> 99501f174e58

Successfully built 99501f174e58
Successfully tagged reizio_reiz:latest
Building web
Step 1/5 : FROM python:3.8.2-slim
 ---> e7d894e42148
Step 2/5 : RUN apt-get update  && apt-get upgrade -y  && apt-get install git apt-utils -y
 ---> Using cache
 ---> e065aae1ac6d
Step 3/5 : RUN git clone https://github.com/treescience/search.tree.science
 ---> Using cache
 ---> 35db27c91c06
Step 4/5 : WORKDIR search.tree.science
 ---> Using cache
 ---> 4e636879d5de
Step 5/5 : ENTRYPOINT ["python", "-m", "http.server", "8080"]
 ---> Using cache
 ---> 9ef76d9cae2a

Successfully built 9ef76d9cae2a
Successfully tagged reizio_web:latest
Starting reizio_edgedb_1 ... done
Recreating reizio_reiz_1 ... done
Recreating reizio_web_1  ... done
Attaching to reizio_edgedb_1, reizio_reiz_1, reizio_web_1
edgedb_1  | WARNING: no logs are available with the 'none' log driver
reiz_1    | + mkdir -p /app/tmp/data
reiz_1    | + mkdir -p /root/.local/
reiz_1    | + cp static/configs/docker_config.json /root/.local/reiz.json
reiz_1    | + pre_db
reiz_1    | + python -m reiz.sampling.get_dataset --limit 10 /app/tmp/dataset.json
reiz_1    | [2021-05-24 05:18:11,043] get_pypi_dataset --- Adding six to the dataset
reiz_1    | [2021-05-24 05:18:11,056] get_pypi_dataset --- Adding urllib3 to the dataset
reiz_1    | [2021-05-24 05:18:11,060] get_pypi_dataset --- Adding requests to the dataset
reiz_1    | [2021-05-24 05:18:11,066] get_pypi_dataset --- Adding certifi to the dataset
reiz_1    | [2021-05-24 05:18:11,074] get_pypi_dataset --- Adding setuptools to the dataset
reiz_1    | [2021-05-24 05:18:11,076] get_pypi_dataset --- Adding idna to the dataset
reiz_1    | [2021-05-24 05:18:11,099] get_pypi_dataset --- Adding botocore to the dataset
reiz_1    | [2021-05-24 05:18:11,101] get_pypi_dataset --- Adding chardet to the dataset
reiz_1    | [2021-05-24 05:18:11,104] get_pypi_dataset --- Adding s3transfer to the dataset
reiz_1    | [2021-05-24 05:18:11,113] get_pypi_dataset --- Adding pip to the dataset
reiz_1    | [2021-05-24 05:18:11,168] get_pypi_dataset --- 10 repositories have been added to the /app/tmp/dataset.json
reiz_1    | + python -m reiz.sampling.fetch_dataset /app/tmp/dataset.json /app/tmp/rawdata
reiz_1    | [2021-05-24 05:18:11,973] fetch           --- 'six' has been checked at 3974f0c4f6700a5821b451abddff8b3ba6b2a04f revision
reiz_1    | [2021-05-24 05:18:11,974] fetch           --- 'certifi' has been checked at c5311ad0533e78240f34b0e35451794d0e261d54 revision
reiz_1    | [2021-05-24 05:18:12,073] fetch           --- 'urllib3' has been checked at 97a16d74f287ce84dcb14aa90bf28c9088579257 revision
reiz_1    | [2021-05-24 05:18:12,123] fetch           --- 'requests' has been checked at f6d43b03fbb9a1e75ed63a9aa15738a8fce99b50 revision
reiz_1    | [2021-05-24 05:18:12,249] fetch           --- 'idna' has been checked at b0ef4bf7fe79174e9ddd8d89e2f48b2e6dc3e721 revision
reiz_1    | [2021-05-24 05:18:12,576] fetch           --- 's3transfer' has been checked at 59e968d05288092948284001710c416677102266 revision
reiz_1    | [2021-05-24 05:18:12,603] fetch           --- 'setuptools' has been checked at a5131f0b82e098da6c07a03a47f36f3a52f73fb6 revision
reiz_1    | [2021-05-24 05:18:12,706] fetch           --- 'chardet' has been checked at e6488640addf2af696e1d602e9007d9720d9fd7d revision
reiz_1    | [2021-05-24 05:18:13,493] fetch           --- 'pip' has been checked at 0a49dd913f673b8abb7c597d58f0e26d24dc34bc revision
reiz_1    | [2021-05-24 05:18:18,498] fetch           --- 'botocore' has been checked at b9cdb88f404c67acfb7088dbbbe97279baa1aa88 revision
reiz_1    | + python -m reiz.sampling.sanitize_dataset /app/tmp/dataset.json /app/tmp/rawdata /app/tmp/data --ignore-tests
reiz_1    | [2021-05-24 05:18:18,713] sanitize_dataset --- 'six' has been sanitized
reiz_1    | [2021-05-24 05:18:18,714] sanitize_dataset --- 'idna' has been sanitized
reiz_1    | [2021-05-24 05:18:18,714] sanitize_dataset --- 'certifi' has been sanitized
reiz_1    | [2021-05-24 05:18:18,714] sanitize_dataset --- 's3transfer' has been sanitized
reiz_1    | [2021-05-24 05:18:18,714] sanitize_dataset --- 'requests' has been sanitized
reiz_1    | [2021-05-24 05:18:18,714] sanitize_dataset --- 'urllib3' has been sanitized
reiz_1    | [2021-05-24 05:18:18,715] sanitize_dataset --- 'chardet' has been sanitized
reiz_1    | [2021-05-24 05:18:18,715] sanitize_dataset --- 'setuptools' has been sanitized
reiz_1    | [2021-05-24 05:18:18,715] sanitize_dataset --- 'pip' has been sanitized
reiz_1    | [2021-05-24 05:18:18,715] sanitize_dataset --- 'botocore' has been sanitized
reiz_1    | + sleep 15
reiz_1    | + post_db
reiz_1    | + python /app/scripts/create_db.py
reiz_1    | [2021-05-24 05:18:34,006] create_db       --- database exits, doing nothing...
reiz_1    | + python -m reiz.serialization.insert --fast --limit 75 --project-limit 7 /app/tmp/dataset.json
reiz_1    | + python -m reiz.web.api
reiz_1    | [2021-05-24 05:18:35 +0000] [140] [INFO] Goin' Fast @ http://0.0.0.0:8000
reiz_1    | [2021-05-24 05:18:35,211] _helper         --- Goin' Fast @ http://0.0.0.0:8000
reiz_1    | [2021-05-24 05:18:35 +0000] [140] [INFO] Starting worker [140]
reiz_1    | [2021-05-24 05:18:35,482] serve           --- Starting worker [140]
```

The search engine can be visited at 8080 port. However, no new log is printed and no search result is ever returned:
![image](https://user-images.githubusercontent.com/6076043/119300826-405bea00-bc16-11eb-8c80-6a5424f27687.png)
