### 2021-05-24 #3296 @yuhc
> Interesting, I can successfully spin the instances with 3.9 though will definitely investigate (would you mind opening an issue on the tracker).

Created https://github.com/reizio/reiz.io/issues/51.

> You should wait the API to start before accessing, since without it the web ui will just wait and timeout eventually. Would you mind sending me the logs (btw it would be better if you could create an issue on the repo itself!)

Created https://github.com/reizio/reiz.io/issues/52.
