### 2021-05-24 bec1993
serialization: calculate the max_active_tasks properly
