### 2021-05-24 a31824c
docker: use 3.8 instead of 3.9
