### 2021-05-24 reizio/reiz.io#51 @isidentical
It is indeed interesting that the matrix doesn't contain `3.9` but the examples in the same page use it; https://docs.docker.com/compose/compose-file/compose-file-v3/#build

I fixed this on the upstream, thanks for the report!
