### 2021-05-26 reizio/reiz.io#52 @yuhc
I do see more outputs now, but still the search doesn't return.

```
reiz_1    | + python /app/scripts/create_db.py
reiz_1    | [2021-05-26 03:27:15,918] create_db       --- database exits, doing nothing...
reiz_1    | + python -m reiz.serialization.insert --fast --limit 75 --project-limit 7 /app/tmp/dataset.json
reiz_1    | [2021-05-26 03:27:25,130] insert_file     --- 'six/setup.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:25,132] _execute_tasks  --- six: Statistics({Insertion.SKIPPED: 2, Insertion.INSERTED: 1})
reiz_1    | [2021-05-26 03:27:26,566] insert_file     --- 'chardet/setup.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:26,779] insert_file     --- 's3transfer/setup.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:26,900] insert_file     --- 'certifi/setup.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:30,311] insert_file     --- 'idna/setup.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:32,474] insert_file     --- 'botocore/setup.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:32,918] insert_file     --- 'setuptools/setup.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:33,464] insert_file     --- 'certifi/certifi/core.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:34,147] insert_file     --- 'pip/setup.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:34,712] insert_file     --- 'urllib3/setup.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:35,059] insert_file     --- 'setuptools/bootstrap.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:35,335] insert_file     --- 'requests/setup.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:35,709] insert_file     --- 'certifi/certifi/__main__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:36,101] insert_file     --- 'certifi/certifi/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:36,324] insert_file     --- 'certifi/certifi/tests/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:36,325] _execute_tasks  --- certifi: Statistics({Insertion.INSERTED: 5})
reiz_1    | [2021-05-26 03:27:38,025] insert_file     --- 'pip/tools/tox_pip.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:38,243] insert_file     --- 's3transfer/s3transfer/compat.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:38,976] insert_file     --- 'idna/tools/intranges.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:39,486] insert_file     --- 'idna/idna/compat.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:39,811] insert_file     --- 'idna/idna/package_data.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:40,074] insert_file     --- 'botocore/botocore/translate.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:40,658] insert_file     --- 's3transfer/s3transfer/exceptions.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:41,508] insert_file     --- 'pip/tools/release/check_version.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:41,618] insert_file     --- 'setuptools/pavement.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:41,708] insert_file     --- 'idna/idna/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:43,829] insert_file     --- 'setuptools/conftest.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:44,151] insert_file     --- 's3transfer/s3transfer/subscribers.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:45,091] insert_file     --- 's3transfer/s3transfer/constants.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:45,261] insert_file     --- 'setuptools/tools/ppc64le-patch.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:45,994] insert_file     --- 'botocore/botocore/history.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:46,008] insert_file     --- 'requests/docs/_themes/flask_theme_support.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:48,294] insert_file     --- 's3transfer/s3transfer/delete.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:49,988] insert_file     --- 'chardet/bench.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:50,851] insert_file     --- 'setuptools/tools/finalize.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:51,249] insert_file     --- 'idna/idna/codec.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:51,473] insert_file     --- 'urllib3/noxfile.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:51,577] insert_file     --- 'requests/requests/compat.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:53,238] insert_file     --- 'idna/idna/intranges.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:53,239] _execute_tasks  --- idna: Statistics({Insertion.INSERTED: 7, Insertion.SKIPPED: 3})
reiz_1    | [2021-05-26 03:27:53,625] insert_file     --- 'botocore/botocore/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:54,888] insert_file     --- 'setuptools/setuptools/extension.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:54,888] _execute_tasks  --- setuptools: Statistics({Insertion.INSERTED: 7})
reiz_1    | [2021-05-26 03:27:55,977] insert_file     --- 'requests/requests/structures.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:56,530] insert_file     --- 'requests/requests/__version__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:56,832] insert_file     --- 'urllib3/docs/conf.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:57,587] insert_file     --- 'botocore/botocore/errorfactory.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:58,182] insert_file     --- 's3transfer/tests/integration/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:58,183] _execute_tasks  --- s3transfer: Statistics({Insertion.SKIPPED: 12, Insertion.INSERTED: 7})
reiz_1    | [2021-05-26 03:27:58,278] insert_file     --- 's3transfer/tests/unit/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:58,382] insert_file     --- 's3transfer/tests/functional/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:58,383] _execute_tasks  --- s3transfer: Statistics({Insertion.SKIPPED: 12, Insertion.CACHED: 7, Insertion.INSERTED: 2})
reiz_1    | [2021-05-26 03:27:58,493] insert_file     --- 'idna/tests/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:27:58,494] _execute_tasks  --- idna: Statistics({Insertion.CACHED: 7, Insertion.SKIPPED: 3, Insertion.INSERTED: 1})
reiz_1    | [2021-05-26 03:27:59,629] insert_file     --- 'botocore/botocore/retries/special.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:01,690] insert_file     --- 'urllib3/test/benchmark.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:02,604] insert_file     --- 'botocore/botocore/retries/throttling.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:02,605] _execute_tasks  --- botocore: Statistics({Insertion.SKIPPED: 32, Insertion.INSERTED: 7})
reiz_1    | [2021-05-26 03:28:03,091] insert_file     --- 'pip/tools/release/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:04,735] insert_file     --- 'requests/requests/help.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:05,433] insert_file     --- 'chardet/test.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:05,712] insert_file     --- 'requests/requests/packages.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:05,712] _execute_tasks  --- requests: Statistics({Insertion.INSERTED: 7, Insertion.SKIPPED: 2})
reiz_1    | [2021-05-26 03:28:05,714] _execute_tasks  --- certifi: Statistics({Insertion.CACHED: 5})
reiz_1    | [2021-05-26 03:28:05,715] _execute_tasks  --- six: Statistics({Insertion.SKIPPED: 2, Insertion.CACHED: 1})
reiz_1    | [2021-05-26 03:28:06,021] insert_file     --- 'botocore/botocore/retries/base.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:09,051] insert_file     --- 'chardet/chardet/enums.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:11,973] insert_file     --- 'pip/docs/html/conf.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:12,784] insert_file     --- 'urllib3/test/conftest.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:12,941] insert_file     --- 'pip/src/pip/__main__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:13,439] insert_file     --- 'pip/src/pip/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:13,440] _execute_tasks  --- pip: Statistics({Insertion.INSERTED: 7, Insertion.SKIPPED: 2})
reiz_1    | [2021-05-26 03:28:13,615] insert_file     --- 'setuptools/setuptools/errors.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:13,754] insert_file     --- 'setuptools/setuptools/_deprecation_warning.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:14,325] insert_file     --- 'botocore/botocore/retries/adaptive.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:14,931] insert_file     --- 'urllib3/test/tz_stub.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:16,840] insert_file     --- 'urllib3/test/port_helpers.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:16,841] _execute_tasks  --- urllib3: Statistics({Insertion.INSERTED: 7, Insertion.SKIPPED: 1})
reiz_1    | [2021-05-26 03:28:17,635] insert_file     --- 'chardet/chardet/sjisprober.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:19,054] insert_file     --- 'setuptools/setuptools/_imp.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:19,077] insert_file     --- 'chardet/chardet/sbcsgroupprober.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:19,506] insert_file     --- 'chardet/chardet/johabprober.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:19,506] _execute_tasks  --- chardet: Statistics({Insertion.INSERTED: 7, Insertion.SKIPPED: 2})
reiz_1    | [2021-05-26 03:28:19,672] insert_file     --- 'botocore/botocore/retries/bucket.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:21,192] insert_file     --- 'botocore/botocore/retries/quota.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:21,621] insert_file     --- 'botocore/botocore/retries/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:21,988] insert_file     --- 'chardet/chardet/mbcharsetprober.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:22,914] insert_file     --- 'setuptools/setuptools/dep_util.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:22,994] insert_file     --- 'pip/src/pip/_vendor/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:23,981] insert_file     --- 'pip/src/pip/_vendor/packaging/__about__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:27,035] insert_file     --- 'botocore/botocore/docs/waiter.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:27,685] insert_file     --- 'chardet/chardet/charsetgroupprober.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:28,137] insert_file     --- 'chardet/chardet/euckrprober.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:28,183] insert_file     --- 'setuptools/setuptools/namespaces.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:28,608] insert_file     --- 'chardet/chardet/big5prober.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:29,649] insert_file     --- 'chardet/chardet/mbcsgroupprober.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:30,089] insert_file     --- 'chardet/chardet/cp949prober.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:31,036] insert_file     --- 'setuptools/setuptools/lib2to3_ex.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:31,342] insert_file     --- 'setuptools/setuptools/py34compat.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:31,343] _execute_tasks  --- setuptools: Statistics({Insertion.CACHED: 7, Insertion.INSERTED: 7, Insertion.SKIPPED: 5, None: 1})
reiz_1    | [2021-05-26 03:28:31,346] _execute_tasks  --- s3transfer: Statistics({Insertion.SKIPPED: 12, Insertion.CACHED: 9})
reiz_1    | [2021-05-26 03:28:31,349] _execute_tasks  --- idna: Statistics({Insertion.CACHED: 8, Insertion.SKIPPED: 3})
reiz_1    | [2021-05-26 03:28:31,448] insert_file     --- 'urllib3/test/with_dummyserver/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:31,916] insert_file     --- 'urllib3/test/contrib/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:32,558] insert_file     --- 'pip/src/pip/_vendor/packaging/requirements.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:32,665] insert_file     --- 'urllib3/src/urllib3/_compat.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:34,356] insert_file     --- 'botocore/botocore/docs/shape.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:34,356] _execute_tasks  --- botocore: Statistics({Insertion.SKIPPED: 35, Insertion.CACHED: 7, Insertion.INSERTED: 7})
reiz_1    | [2021-05-26 03:28:34,430] insert_file     --- 'pip/src/pip/_vendor/packaging/_compat.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:34,477] insert_file     --- 'chardet/chardet/escprober.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:35,411] insert_file     --- 'urllib3/src/urllib3/filepost.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:35,527] insert_file     --- 'urllib3/src/urllib3/_version.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:36,160] insert_file     --- 'pip/src/pip/_vendor/packaging/_structures.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:39,510] insert_file     --- 'urllib3/src/urllib3/__init__.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:40,828] insert_file     --- 'pip/src/pip/_vendor/packaging/utils.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:41,262] insert_file     --- 'pip/src/pip/_vendor/packaging/_typing.py' has been inserted successfully
reiz_1    | [2021-05-26 03:28:43,934] insert_file     --- 'urllib3/src/urllib3/util/ssl_match_hostname.py' has been inserted successfully
reiz_1    | + python -m reiz.web.api
reiz_1    | [2021-05-26 03:28:44 +0000] [149] [INFO] Goin' Fast @ http://0.0.0.0:8000
reiz_1    | [2021-05-26 03:28:44,831] _helper         --- Goin' Fast @ http://0.0.0.0:8000
reiz_1    | [2021-05-26 03:28:45 +0000] [149] [INFO] Starting worker [149]
reiz_1    | [2021-05-26 03:28:45,117] serve           --- Starting worker [149]
```
