### 2021-05-26 reizio/reiz.io#52 @isidentical
Can you go to `http://localhost:8000/stats` and confirm the API is up and running with real data? You are expected to see something like that (numbers can change);
```json
{"status":"success","results":{"Module":76,"AST":10743,"stmt":2250,"expr":8345},"exception":null}
```
