### 2021-05-27 reizio/reiz.io#52 @yuhc
Yes, the status is actually success

```json
{"status":"success","results":{"Module":187,"AST":26510,"stmt":5701,"expr":20406},"exception":null}
```

However, after the search got stuck, I refreshed the web and re-search using a different query but the stats info didn't change.
