### 2021-05-27 reizio/reiz.io#52 @isidentical
> However, after the search got stuck, I refreshed the web and re-search using a different query but the stats info didn't change.

Can you check network tab on the browser? The stats aren't expected to change since it only contains the number of nodes indexed in the database.
