### 2021-05-27 reizio/reiz.io#52 @isidentical
here are some examples;
![image](https://user-images.githubusercontent.com/47358913/119783433-dec9a300-bed5-11eb-8470-ed4efc2bac68.png)
![image](https://user-images.githubusercontent.com/47358913/119783469-e8530b00-bed5-11eb-9970-accf3120ebb8.png)
