### 2021-05-27 #3296 @lutzhamel
@isidentical, docker-compose does not run on the given files...
```
ubuntu@ip-172-31-94-52:~$ git clone https://github.com/reizio/reiz.io
Cloning into 'reiz.io'...
remote: Enumerating objects: 1884, done.
remote: Total 1884 (delta 206), reused 264 (delta 132), pack-reused 1522
Receiving objects: 100% (1884/1884), 586.58 KiB | 24.44 MiB/s, done.
Resolving deltas: 100% (1098/1098), done.
ubuntu@ip-172-31-94-52:~$ ls
reiz.io
ubuntu@ip-172-31-94-52:~$ cd reiz.io/
ubuntu@ip-172-31-94-52:~/reiz.io$ docker-compose up --build --remove-orphans
ERROR: Version in "./docker-compose.yml" is unsupported. You might be seeing this error because you're using the wrong Compose file version. Either specify a supported version (e.g "2.2" or "3.3") and place your service definitions under the `services` key, or omit the `version` key and place your service definitions at the root of the file to use version 1.
For more on the Compose file format versions, see https://docs.docker.com/compose/compose-file/
ubuntu@ip-172-31-94-52:~/reiz.io$
```
