### 2021-05-27 reizio/reiz.io#52 @yuhc
Hi @isidentical , here are a few screenshots:

8000/stats:
![image](https://user-images.githubusercontent.com/6076043/119866623-d707ff80-bed1-11eb-90df-cd93e3efd9e4.png)
![image](https://user-images.githubusercontent.com/6076043/119867066-51d11a80-bed2-11eb-8187-cf169dc57d9a.png)

8080:
![image](https://user-images.githubusercontent.com/6076043/119867104-60b7cd00-bed2-11eb-8967-8b4db2020723.png)
