### 2021-05-27 #3296 @lutzhamel
Here is what I am using:
```
ubuntu@ip-172-31-94-52:~/reiz.io$ docker --version
Docker version 20.10.6, build 370c289
ubuntu@ip-172-31-94-52:~/reiz.io$ docker-compose --version
docker-compose version 1.25.0, build unknown
ubuntu@ip-172-31-94-52:~/reiz.io$
```
