### 2021-05-27 #3296 @yuhc
Hi @isidentical , I just uploaded a screenshot of the 8000 page to https://github.com/reizio/reiz.io/issues/52, and hope it would be helpful for your debugging. While I don't think the review is going in the right direction and I think I should state it clearly in case you may not know:

1. The repo, website or article proof should contain the running environment of your software. You should either support the code for multiple environments, or state the supported environments clearly.
    1. I'm using AWS c4.4xlarge instance, Ubuntu 20.04 and the latest packages you can get on 20.04 by default to run the code. IDK what settings @lutzhamel is using, but the expected testing environment **should** be provided by you.
    2. Otherwise, you should get a c4 instance and fix your code to work there. You can rent a c4.xlarge for a few hours and it will just cost you a buck.
1. The reviewers aren't responsible for helping with the debug.
    1. This is a basic criterion for any peer-review journal or conference, and I believe it also applies to JOSS.
    2. From now on I'll wait for the setup results from @lutzhamel and updated setting instructions from you, otherwise the process will be a Pandora's box for the reviewers.
