### 2021-05-27 #3296 @isidentical
> @isidentical, have you considered setting up a working instance on a web based virtual machine like https://replit.com/ eliminating set up issues all together?

That is a great idea! i'll create an open instance as well as the instructions on a clear aws machine. Sorry for all the inconvenience i caused!
