### 2021-05-30 4dfa8f3
web: integrate the web UI to simplify installation
