### 2021-06-01 692c824
serialization: handle config values better (allow options to default when they are None)
