### 2021-06-02 #3296 @isidentical
@yuhc @lutzhamel I've deployed a public instance on the web address: `https://reiz.io`. Unfortunately, I wasn't able to give exact instructions for an AWS instance since they still haven't approved my personal account though as you stated, if any of you want to have access to the environment the server is running I can give access (it is running on a digital ocean VPS right now). Please let me know if this method works for you or not, thanks!
