### 2021-06-05 #3296 @yuhc
Thanks for setting up reiz.io. It works for me.
Could you update the docs to include your exact setup steps on DigitalOcean? I don't think it really matters which cloud service you choose. The main reason why we asked you to install reiz.io on a clean VM is that we want a reproducible instruction.

BTW, registering an AWS account should just take a few minutes. I guess something went wrong, and you may need to contact the support.
