### 2021-06-05 #3296 @isidentical
>  Could you update the docs to include your exact setup steps on DigitalOcean? I don't think it really matters which cloud service you choose. The main reason why we asked you to install reiz.io on a clean VM is that we want a reproducible instruction.

Sure! I think the `docker-compose` should just work in any case (the errors should have gone by now (it was a browser thing as far as I can locate with the confusion that the API is served on another port and the website is another port), since I've also made a couple changes for supporting port forwarding over ssh) but I'll double check it. As far as the reproducibility goes, I'll also include a document to install it in the deploy mode (docker is for creating a toy environment with limited data, but this installation method installs ~500 projects unlike 10 on docker).
