### 2021-06-08 cff3cc6
docs: add installation steps on AWS
