### 2021-06-08 #3296 @isidentical
My AWS account is also now got approved, so here is the exact steps to use `reiz` with only `docker`/`docker-compose` and ssh; https://reizio.readthedocs.io/en/latest/installation-to-aws.html
