### 2021-06-13 #3296 @yuhc
Hi @isidentical-
I took a look at the updated README, and it looks good in general. There are some nits:
* The login section can be removed. Port forwarding is actually not necessary or recommended.
* `git fetch` and `git reset --hard` lines can be removed.
* User should expect to visit reiz via `instance_ip:8000` instead of `localhost:8000`.
** Or you can remove anything related to AWS and change the instruction to the one for local Ubuntu 18.04.
