### 2021-06-13 #3296 @yuhc
Hi @mjsottile, the article proof link is broken. Could you help fix it?
