### 2021-06-13 #3296 @yuhc
While waiting for the article proof, I have a question for you, @isidentical:
The software is about code search **at scale**, but the database setup seems to take a long time and there's no experiment showing the query latency for scaled datasets. Is there any evidence with which you claim the search system to be "at scale"? (Something like algorithms, complexity analysis would be fine. Just for reference.)
