### 2021-06-13 #3296 @isidentical
> but the database setup seems to take a long time

Yes, indeed. Reiz does multiple phases while aggregating data to the database. This spans from the actual downloading of the projects to generate a one-time [node-tag](https://github.com/reizio/reiz.io/blob/cff3cc6eaad532ac1a956c1f7c7a58d97ea00e4b/reiz/serialization/transformers.py#L48-L69) (a fingerprint of the node) and since all of these have to be done in the setup/indexing part, neither users nor the running system has to pay the cost after installation (on the actual run).

> there's no experiment showing the query latency for scaled datasets.

I'm not sure where to put it (like in paper or the software docs), but yes definitely I can index data at the real-world level (50k+ unique files from ~1000 projects) and let you know about the latency (which would probably be performed directly over the server with the API, to not effected by other factors). I have a few examples where Reiz is used in the development of CPython (to collect data, e.g whether to deprecate a usage or not / checking out how many arguments have annotations separated across lines for a new feature), and I intend to use those queries in the benchmarks (let me know if you have any extra queries that you want me to add to that list). It would be a bit hard to compare it with other providers (just like we did at the evaluation), though it might help to validate the point of `at scale`.

> (Something like algorithms, complexity analysis would be fine. Just for reference.)

`at scale` term originated from the fact that for other tools that tried to do the same (search source code directly on AST, but not as in the manner of 'search engine') performed worse compared to Reiz when you'd run them on a giant dataset (50k files for example).

There is a small 'internal use' utility under Reiz called `./scripts/debug_query.py` which prints out the AST and IR for the given query (it requires to be installed in full mode, with the configuration being set). It is helpful when you want to see the complexity of a single query. Here is a [simple example](https://gist.github.com/isidentical/820863d5e97576c8bc06ca4fd8fefc28), and [here is the](https://gist.github.com/isidentical/d29ccd75a46200305b01dd62db57cc48) generated IR of the most complex query that I have. The complexity increases by the number of 'checks' you are performing. I don't think it would make sense to put this into the scale of big-O, though abstract implementation tries to be as linear as possible. It used to create a new block every time you used some of the advanced matches, but now they are all flattened out thanks to https://github.com/reizio/reiz.io/pull/12.
