### 2021-06-13 #3296 @yuhc
Seems there are some misunderstandings of "at scale". Reiz is a good tool, and I just wanted to know whether its title is a bit overstated.

> This spans from the actual downloading of the projects to generate a one-time [node-tag](https://github.com/reizio/reiz.io/blob/cff3cc6eaad532ac1a956c1f7c7a58d97ea00e4b/reiz/serialization/transformers.py#L48-L69)

That's the concern. How long does it take to generate a dataset for 1k projects, 100k, 1M, 10M...? 50k files, 500k, 5M, 50M...? 50k files and 1k projects are of small scale to be honest. In most cases, the generation/update also has to be done on-line for a search engine, but that is out of scope.

> I'm not sure where to put it (like in paper or the software docs)

Most framework- or system-level software has performance analysis reports in their documentation (for example, a "Performance" page in https://reizio.readthedocs.io/en/latest/reizql.html). I will not doubt its importance, but I do not force you to measure the performance at this moment as reiz is not yet a production-ready system. I am pretty sure you will need it in the future.

> but yes definitely I can index data at the real-world level (50k+ unique files from ~1000 projects) and let you know about the latency (which would probably be performed directly over the server with the API, to not effected by other factors).

As I said, this data scale is not really impressive. And please remember, there is not only single access to reiz: thousands or tens of thousands requests could be sent simultaneously. I will not call a system "at scale" if it cannot prove that its ability to handle a significant size of datasets and requests.

Why I asked for latency: If there are measurements for 50k, 500k, 5m files and 100, 1k, 10k requests, and they show similar latencies, reiz then proves that it is scalable.

> I have a few examples where Reiz is used in the development of CPython

CPython, so what? It is a great tool, but I will not say my system is scalable because it is written with specific language or compiler.

> `at scale` term originated from the fact that for other tools that tried to do the same (search source code directly on AST, but not as in the manner of 'search engine') performed worse compared to Reiz when you'd run them on a giant dataset (50k files for example).

That is "efficiently" instead of "at scale". There is a simple question that you may think about: **assume** reiz can support indexing C/C++/Java files, could it run on a Linux kernel dataset efficiently? Chromium? AOSP?
Again, without any perf numbers, a potential user of reiz would only try reiz with small-sized personal projects. "At scale" is meaningless in such a case.
