### 2021-06-13 #3296 @isidentical
Fair points! I wasn't imagining on those points, and my comparison was with limited tools. I'll try to update the docs with benchmarks on the data that I can easily collect through the already integrated sampling process, and also drop the `at scale` tag at the end. Thanks!
