### 2021-06-14 #3296 @lutzhamel
Hello:

I have some comments on the current state of the manuscript/software:

1. The documentation does not explain how to set up the software package for a user specific project, GitHub or otherwise.  It only demonstrates the functionality on some prepackaged set of projects.
2. The query language feels rough and the syntax is unwieldy, creating a huge barrier of entry to the package.  Basically the user is asked to construct AST snippets in the precise format that CPython specifies that act as templates against the DB.  I am wondering if query languages for graph databases can give some inspiration for an easier query syntax.
3. Finally, why not incorporate regular expressions into the `match_string_pattern`?

@mjsottile, given @yuhc comments and given my comments above, I feel that the software is not mature enough at this point to warrant publication.  I feel that the software needs to be deployed in some real-life settings and demonstrate benefits before the paper can be accepted.
