### 2021-06-14 #3296 @isidentical
Hello @lutzhamel!

> The documentation does not explain how to set up the software package for a user specific project, GitHub or otherwise. It only demonstrates the functionality on some prepackaged set of projects.

Yes, that was intentional. Even though this is very trivial to do so, and possible (any package hosted on some sort remote git repo can be used, through https://github.com/reizio/reiz.io/blob/cff3cc6eaad532ac1a956c1f7c7a58d97ea00e4b/reiz/sampling/data.py#L8-L19, which corresponds to the `dataset.json` file entries) I'd avoided this since i wanted to keep Reiz as a general purpose search engine. Reiz also employs the design of pluggable-parts, so the implementors (in this case reiz.io) provides the sampling strategies. The default strategy is fetching most downloaded packages from the Python's package index, PyPI though there have been others (even though not turned on by default), such as [fetching most popular projects from github by either the number of stars or the last update](https://github.com/reizio/source/blob/master/source/providers/github.py).

> The query language feels rough and the syntax is unwieldy, creating a huge barrier of entry to the package. Basically the user is asked to construct AST snippets in the precise format that CPython specifies that act as templates against the DB. I am wondering if query languages for graph databases can give some inspiration for an easier query syntax.

I'd have to disagree with you on this point. The language is entirely free from the CPython or whatever implementation's AST (as described in the Grammar appendix). When you are implementing a variant for Reiz, you feed an ASDL file which then creates those 'matchers'. In the reference implementation I proposed, the ASDL file originated from CPython ([see this](https://github.com/reizio/reiz.io/blob/master/static/Python-reiz.asdl)). If you were to implement another language this would be totally different.

However I agree that ReizQL as is a low level language for querying source code, and that is also intentional. I even wrote a high level interface for Python called [irun](https://github.com/reizio/irun). The following query is written in a python superset where you are searching code that looks exactly the same with some expanded fragments which is then compiled to the ReizQL

Query:
```py
with open(...) as $stream:
    tree = ast.parse($stream.read())
```

ReizQL version:
```
With(
    items=[
        withitem(
            context_expr=Call(func=Name(id="open"), args=[...], keywords=[]),
            optional_vars=~stream,
        )
    ],
    body=[
        Assign(
            targets=[Name(id="tree")],
            value=Call(
                func=Attribute(value=Name(id="ast"), attr="parse"),
                args=[
                    Call(
                        func=Attribute(value=~stream, attr="read"), args=[], keywords=[]
                    )
                ],
                keywords=[],
            ),
        )
    ],
)
```

As you said, writing the second form by hand would create a really high level of entry. Though as I tried to explain, Reiz is the just the execution engine. This is merely a syntactical sugar on top of it.

> Finally, why not incorporate regular expressions into the match_string_pattern?

When you meant to match some names (e.g finding all tests written in pytest/unittest fashion, `FunctionDef(f'test_%')`) you often search for prefixes or suffixes. Of course I don't deny the need of having something rather complicated, but in terms of execution cost that would be very expensive compared to a very stricter but still powerful alternative (LIKE syntax in SQL). Since the underlying engine have native support for incorporating regexes, one day, when the need arises it would be possible to implement this fairly easily, though some implementations might need to validate / check the complexity of input regexes or otherwise it might be performing really bad since we don't have trigram indexes unlike other engines that support it.

>  @mjsottile, given @yuhc comments and given my comments above, I feel that the software is not mature enough at this point to warrant publication. I feel that the software needs to be deployed in some real-life settings and demonstrate benefits before the paper can be accepted.

Even though that saddens me, I do understand the source of worries. Thank you both for your reviews/suggestions.
