### 2021-06-14 #3296 @yuhc
Hi @isidentical, I think @lutzhamel made a really good point. When you advertise your software, you need to document its features and probably also limitations well. Any undocumented feature is assumed not to exist, while anything implied from the context but undocumented in the limitations is supposed to be supported.

From the article proof and repo, we expect to see the instructions and examples for an **efficient** (scalable, previously), **user-friendly** code search on a **user-specific** project. Your document should demonstrate these points or admit some shortcomings.
I've talked about the efficiency for which no data is provided. @lutzhamel mentioned user-friendliness and generality of the software, and I'd like to put a question onto your plate more explicitly:
* When would we want to write tens of lines of ReizQL to search code with Reiz, instead of using a single line of regexpr with any existing code search engine that supports fuzzy search?

I'd also like to mention that AST-based code search and programming, or more generally code-to-code code search and programming, have been researched for decades. There are plenty of papers that you can find and read to improve Reiz and its documentation.
