### 2021-06-15 #3296 @mjsottile
Hi all - I’ve been a bit buried in day job activities so I’ve been watching the discussion from the sidelines.  Thank you for the detailed discussion.

Regarding the review itself:

- Remember that the criteria for a JOSS paper are slightly different than would be applied to a full journal or conference paper.  JOSS papers are typically quite short and do not have space for extensive evaluations of performance, usability, and so on.  For reference, the criteria are here : https://joss.readthedocs.io/en/latest/review_criteria.html .
- There is no requirement that JOSS papers be deployed in "real-world" settings or in any production-class environment for acceptance.  They must reflect scholarly work of relevance to the research community.  See the same link for our definition of “substantial scholarly effort”.
- Research artifacts are often works in progress (there is *always* more to do given time and resources).  We don't gate acceptance on having an empty todo list, so even if there might be something useful to have (e.g., regex as part of a query), that's not a strong justification to not accept.
- That said, I am inclined to think that publishing this paper in JOSS is premature since typically JOSS publishes artifacts related to existing research efforts.  JOSS papers are typically too small to sufficiently describe the research work itself.  I think that since Reiz is a research project in and of itself, and not just a tool used as part of a different research project, it would be worth writing up a proper paper.  That paper would be related to the research effort around the code search method where you have the space to go into detail about how it works, how it relates to prior work, and perform a detailed assessment related to scaling and/or efficiency.  This seems like a perfectly valid submission to the software engineering or programming languages literature.  If such a paper was written that established those details, then a JOSS submission that focuses specifically on the software artifact itself would be much easer to justify as acceptable since it would have the other paper to reference.

Comments to @isidentical [note: I'm sharing these comments for your benefit - they may end up ultimately being useful if you choose to write a full paper for a regular research paper venue]:

- You should be conscious of making sure certain things are not overstated without evidence, such as performance or scaling.  I would consider dropping things like "at scale" from a paper title if scaling isn't clearly demonstrated.  The same goes for claims of efficiency.
- You should improve the statement of need.  This comment thread makes that apparent - the paper makes some mention of the deficiencies of text-oriented search due to the fact that program code represents a structure, not linear text.  You can also see that in the results where alternative search mechanisms have high false positive counts.  Calling that out and explaining WHY Reiz improves on it should be clearer.  The comparison section could also use some additional text explaining the results in the tables to make it clear what you are showing and what the source of false positives are.
- Other parts of the paper need work as well.  I would revise the state of the field, explaining the results in the evaluation/comparison section, and potentially adding the example from https://github.com/openjournals/joss-reviews/issues/3296#issuecomment-860960347 in order to illustrate the query DSL and the tree structure it desugars to.  I also noticed that the references just sort of start without a section heading indicating where the reference section starts and the previous section ends - I would add a section heading in for that.

With respect to some technical details:

- String search and tree search are very different beasts - regexp is appropriate for searching the code as linear text, while tree matching is appropriate for searching the parse tree or AST derived from the text.  The fact this came up in the thread above tells me the paper doesn’t do an adequate job explaining why tree matching is more appropriate for code search.
- The ReizQL mechanism appears similar to other tree search methods in which a template is defined in terms of the node constructors and corresponding subtrees labeled with names holes.  Even if one was to move to something like a generic graph query language to write the pattern, you would still need the query to be aware of the specific types of nodes that could appear in the template.  So the underlying tree representation would bubble up into the query anyways.  The syntactic sugar mentioned in a previous comment (https://github.com/openjournals/joss-reviews/issues/3296#issuecomment-860960347 ) looks like other systems I've seen where a DSL is used to reduce the burden on the query writer to know the specifics of the actual underlying tree representation.  I think the approach taken is pretty reasonable since I've seen the same approach taken in other projects with similar goals.

Any comments/thoughts/questions?
