### 2021-06-15 #3296 @yuhc
It's fine to me to accept Reiz with some minor changes, mainly in its article proof and documentations:

* Rewrite the documentation with what Reiz is able to do **now**, something like: Reiz is an AST-based code search engine which runs ReizQL to search AST patterns in the prebuilt database of Python codes. Feel free to decorate the sentence.
* State the use cases of Reiz as @mjsottile suggested.
* Cite related papers and open-source projects in the documentation.

As @mjsottile mentioned, the comparison between string search and tree search, and using DSL for code query has been well researched. @isidentical, I recommend you to investigate those works more deeply and come up with a nicer README for Reiz. For example, you compared Reiz with GitHub code search which is reasonable at a glance. But if you look closer into GitHub code search, you'd find it's based on ElasticSearch in which one can reimplement a scalable, efficient, extensive "ElasicReiz.io" with trivial effort.

Honestly, the academic value of Reiz is low: AST-based code manipulation is not a popular topic anymore, and alternatives to Reiz appeared many years ago. However, I'm interested in seeing Reiz grow into a more mature project like https://github.com/jonatas/fast and https://github.com/takluyver/astsearch. It can be a useful tool/library for Python once it proves its efficiency comparing with existing libraries. That's why I talked about the **efficiency** so much, and probably why @lutzhamel talked about its programmability. The opportunity that I can see in Reiz is to make it more user-friendly and as fast as possible.
