### 2021-06-15 #3296 @lutzhamel
To add to @yuhc comments.  I am not opposed to publishing this paper as long as the following is fulfilled,
1.  It is made clear that this is the evaluation of an *engine*, that means all the metrics etc that come with evaluating an engine should be included.  See @yuhc comments on this.
2. It is made clear that the query language is not an end user product but it is a prototype AST-based query language to exercise the engine.

Both of these points are not made clear in the current paper and therefore were the source of my confusion.
