### 2021-06-17 #3296 @mjsottile
At this point I believe I have what I need to know from the reviewers @yuhc and @lutzhamel.  I need to think a bit since over the span of a few days the recommendation swung from not acceptable to acceptable.  Two comments though that I believe must be addressed:

> Honestly, the academic value of Reiz is low: AST-based code manipulation is not a popular topic anymore

This is not supported by activity in the literature. I would refrain from dismissing an area of research as part of a review - it is often unproductive and rarely holds up under scrutiny.  A cursory search for recent publications in that relatively broad area yields a steady number of papers up to and including this year.  AST-based code search is still active in areas including but not limited to: mining code repositories, clone detection, defect detection, and is a component of some software synthesis tools.  AST-based code *manipulation* is certainly active - almost all code transformation methods are based on either direct manipulation of the AST or operations on AST-derived structures.

Similarly:

> But if you look closer into GitHub code search, you'd find it's based on ElasticSearch in which one can reimplement a scalable, efficient, extensive "ElasicReiz.io" with trivial effort.

It is also generally not good form to dismiss work in this way.  It is one thing to refer to another tool as potentially having high overlap to encourage an author to discuss what is different or to state that it is basically the same, but implemented using different methods. Casually dismissing it as trivial is not very productive.

Thanks again for the reviews and I'll figure out next steps once I've heard thoughts from @isidentical regarding how they would like to proceed.
