### 2021-06-17 #3296 @yuhc
Thank you @mjsottile for the comments. Yes I brought those up to give some suggestions for the directions of the improvement, but I didn't mean to dismiss the work or the field. I totally agree with what you said and I think we're talking about different scope of "academic values".

AST-based code search was actively explored in the academia so many years ago and I saw quite some amazing papers when I was waiting for Reiz's bug fix. The field isn't that active now not because it's unimportant: The techniques have been used in many places. However, it's because the field has been well explored. There's of course likely to appear more nice works and findings but one should think about it carefully when she/he wants to dive in (for breaking "academic values").

Reiz is an impressive project and I believe it can be useful in cases. (I personally really admire @isidentical for writing this software alone in his age.) The example of ElasticSearch is supposed to tell that it's important to look around and learn from existing works to avoid overstating any features or novelty. Sorry for the confusion.
