### 2021-06-20 64ecc9b
docs, paper: work about suggestions
docs:
- Add a QA section to the docs to mention about source of data
- Add the new performance evaluation section to the docs, also the benchmark script itself
- Add the new related projects section to cite/mention about related projects

paper:
- Drop the 'at scale' suffix from the title
- Mention about Reiz being the execution engine compared to others being public platforms
- Directly show the ReizQL, and also give an example about IRun language
- Change the statement of need to reflect how reiz solves
- Add a conclusion section to the evaluation about wh
[…]
