+++ paper/source/evaluation.md
@@ -3,14 +3,12 @@
 
 For the limited subset of things that can be described in any of the competitor engines, we
 evaluate the performance of `Reiz` by running similiar queries in Github Code Search \[@githubcodesearch\]
-grep.app \[@grepapp\] and Krugle \[@krugle\] and report back the amount of true / false positives.
-
-## Comparisons
-
-The method of evaluation is running the queries and analyzing the top 10 results. Some engines offer
-multiple matches per result, and they will be marked as true positive if any of them are a true
-positive. We also discard match spans, since none of the competitors can successfully display the expression
-boundaries.
+grep.app \[@grepapp\] and Krugle \[@krugle\] and report back the amount of true / false positives. Each
+result that contains an exact match with the intended objective, e.g a call to `len(...)` function will
+be counted as a true positive, and otherwise will be counted as a false positive.  Some engines offer
+multiple matches per result, and they will be marked as true positive if any of them checks out with
+the objective. We also discard match spans, since none of the competitors can successfully display
+the expression boundaries.
 
 Objective: search for a `len(...)` call
 
@@ -42,3 +40,11 @@ Objective: search for a return statement that returns a tuple `return ..., ...`
 | grep.app             | `return ((.*))(,\s*(.*))+`     | 0              | 10              |
 | grep.app (2)         | `return \(((.*))(,\s*(.*))+\)` | 9              | 1               |
 | Reiz                 | `Return(Tuple())`              | 10             | 0               |
+
+
+## Conclusion
+
+On all three objectives, Reiz got all the matches as true-positive due to it's nature of
+being able to leverage syntax trees as well as other annotations that it could collect at
+pre-processing stage (such as node boundaries to report the exact location) unlike others
+where the source code is exercised like a regular textual document.
