### 2021-06-20 #3296 @isidentical
Thanks @yuhc @mjsottile @lutzhamel for your comments! I'm not sure if you still want to proceed or not, though by looking at the last comments I was able to collect a list of stuff to change in the documentation/paper itself and went with them ( https://github.com/reizio/reiz.io/commit/64ecc9b754d3698a328d984fc0f08bf5db21a5f9).

(I don't know how to change the title of this issue, though I dropped the `at Scale` part from the paper itself. )
