### 2021-06-20 #3296 @lutzhamel
OK. That looks pretty good to me.  One major sticking point, and a point that originally got me confused about what I am looking at, remains: The section "Sampling Source Code".  It is written as if the engine can be only used in this context.  This should be rewritten in order to make explicit that this is the experiment in order to validate the engine.  The user is free to index any source code project they want.

With this rewrite I'd be happy to suggest the paper for publication.

-L
