### 2021-06-23 #3296 @yuhc
The only missing piece is that the article doesn't mention existing AST-based search engines or libraries in "State of the field". It looks good to me to move forward after the relevant citations are added.
