+++ paper/paper.bib
@@ -88,3 +88,19 @@
     title        = {EdgeDB: The next generation relational database},
 }
 
+@misc{fast,
+    publisher    = {jonatas},
+    url          = {https://github.com/jonatas/fast},
+    date         = {2017},
+    journaltitle = {GitHub repository},
+    title        = {fast: Find in AST},
+}
+
+@misc{astsearch,
+    publisher    = {takluyver},
+    url          = {https://github.com/takluyver/astsearch},
+    date         = {2014},
+    journaltitle = {GitHub repository},
+    title        = {astsearch: Intelligently search in Python code},
+}
+
