### 2021-06-23 #3296 @isidentical
> The only missing piece is that the article doesn't mention existing AST-based search engines or libraries in "State of the field". It looks good to me to move forward after the relevant citations are added.

Added the fast and the astsearch into the statement of the field section @yuhc (in https://github.com/reizio/reiz.io/commit/873986243b07d11fb6f1a7949c339894bfdabc98, for a diff of changes)
