### 2021-06-24 #3296 @mjsottile
@isidentical : Looks improved.  Some minor edits to correct grammatical mistakes listed by line number in the PDF:

17: "we present Reiz.  A source code search engine" -> "we present Reiz: a is a source code search engine"
23: uses -> use
23: "code is behaved not much different than" -> "code is treated no differently than"
29: "semantical" -> "semantic"
32: "is making them" -> "makes them"
42: "It is a common format" -> "ASDL is a common format"

(Note: may add one sentence at the end of line 43 indicating that the frontend is responsible for instantiating the ASDL terms from whatever AST or parse tree it creates internally for a given language.)

54: "on the runtime" -> "at runtime"
58: capitalize URL
108: "as well as as can be hand written" -> "as well as being hand written"
109: "For example, IRun project" -> "For example, the IRun project"
110: replace semicolon with colon at the end of the line
127: "due to it’s nature of being able to leverage syntax trees" -> "due to it's ability to leverage syntax tree structure"
131: Add section heading "References" separating conclusion from bibliography.

Let me know when you've made those fixes.  It sounds like the reviewers are satisfied with your updated paper and documentation so we can move to the next stages once the corrections are in.
