### 2021-06-24 #3296 @isidentical
> 17: "we present Reiz. A source code search engine" -> "we present Reiz: a is a source code search engine"

There seems to be a typo on this but I wasn't able to infer the correct form. Perhaps `we present Reiz: a source code search engine backend that can exercise queries to match syntax trees ...`?
