### 2021-06-24 1bdd88e
paper: apply editorial suggestions
