### 2021-06-24 #3296 @isidentical
All suggestions should be applied now
