### 2021-06-24 #3296 @mjsottile
Thanks @isidentical.  I don't see any additional work to do, and based on the comments from the reviewers it doesn't appear that any remaining aspects of the submission require attention - even if the check boxes at the top of the issue aren't clicked in the checklists, the topics were addressed in the discussion.

For last steps, please do the following:

- Please create an archive on zenodo or figshare for the submission
- Ensure that the version for the JOSS submission is correct.
- Ensure that the author list in the paper and archive are correct.

Once you've provided that information I can issue the final commands to pass the submission to an EIC for final processing.
