### 2021-06-24 #3296 @isidentical
@mjsottile all points are checked out

> Please create an archive on zenodo or figshare for the submission

https://zenodo.org/record/5029255 (with the DOI `10.5281/zenodo.5029255`)

> Ensure that the version for the JOSS submission is correct.

The `1.0.0` version is correct

> Ensure that the author list in the paper and archive are correct.

The list is correct.
