### 2021-06-25 #3296 @kyleniemeyer
Hi @isidentical, I am doing some final checks for publishing your article.

It looks like most/all of your in-text citations are not using the appropriate citation command. Instead of using `\[` and `\]` to wrap the citations, you should just use `[@author:year]` to let our parser handle the display. The available commands are
```
- `@author:2001`  ->  "Author et al. (2001)"
- `[@author:2001]` -> "(Author et al., 2001)"
- `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
```

I also see some issues with the references at the end to GitHub packages. Could you add the authors for those? It also looks like the "Github code search" reference does not have the correct URL, since it is the same as for "Grep.app".
