### 2021-06-25 #3296 @yuhc
Cool, good work @isidentical !

BTW I'd recommend you to read "Cross-Language Code Search using Static and Dynamic Analyses" (which appears to be the latest work in the AST-based code search area). You'll get some knowledge of the state of the art from it and its citations.
