+++ paper/paper.md
@@ -32,7 +32,7 @@ facing API, maintainers of those libraries do the pre-requisite work of
 collecting samples and estimating the ramifications that operation might cause.
 When the documentation of a framework doesn't sustain the curiosity, searching
 for a structure (e.g a function, a constant) to see how it can be utilized in
-real-world software is a common need among developers \[@developersearch2017\].
+real-world software is a common need among developers [@developersearch2017].
 
 For the problems mentioned above, we present Reiz: a source code search engine
 backend that can exercise queries to match syntax trees in order to leverage
@@ -42,17 +42,17 @@ a `try`/ `finally` construct without knowing what is under the `try` block).
 # State of the field
 
 Popular source code search engines (e.g Github Code Search
-\[@githubcodesearch\]) use full-text search where the code is treated no
+[@githubcodesearch]) use full-text search where the code is treated no
 differently than a regular textual document. Even though this approach works
 for some basic queries, structurally it can't go further than matching token
 sequences. This often causes seeing irrelevant search results on complex
 queries, or even not being able to express the search itself in a purely
 textual form. In the past, there has been some work done regarding making
 queries more expressive through regular expressions (one example might
-codesearch.debian.net \[@debiancodesearch\]), and even annotating the result
+codesearch.debian.net [@debiancodesearch]), and even annotating the result
 set with some semantic and structural knowledge (via finding and resolving API
-names \[@BAJRACHARYA2014241\]). There also have been various tools \[@fast\],
-\[@astsearch\] to search AST patterns within source code, but with a limited
+names [@BAJRACHARYA2014241]). There also have been various tools [@fast],
+[@astsearch] to search AST patterns within source code, but with a limited
 query format and in a file-by-file basis (no index database) which makes them
 quite hard to work with in order to examine a large dataset of source code.
 
@@ -64,11 +64,11 @@ The internals consists of a pipeline that enables the ability to plug in and
 out different components, such as for frontends of different languages. The
 primary piece that every other component directly or indirectly interacts with
 is the Index DB (a.k.a source warehouse) where the serialized AASTs (Annotated
-Abstract Syntax Trees) are being held. It is based on an EdgeDB \[@edgedb\]
+Abstract Syntax Trees) are being held. It is based on an EdgeDB [@edgedb]
 instance which interprets the compiled queries and returns the raw result set.
 The schema used in the Index DB is in the format of ESDL (EdgeDB Schema
 Definition Language) and automatically generated from the host language's ASDL
-\[@asdl97\] declaration. ASDL is a common format used by many different
+[@asdl97] declaration. ASDL is a common format used by many different
 projects, most notably by CPython itself.
 
 ## Sampling Source Code
@@ -76,7 +76,7 @@ projects, most notably by CPython itself.
 Reiz is able to sample source code from a list of projects on a git remote (e.g
 GitHub). The reference implementation for Reiz comes with an indexer which can
 automatically construct a list for the most popular packages on PyPI,
-according to the download statistics \[@hugovankemenade_richardsi_2021\]. The
+according to the download statistics [@hugovankemenade_richardsi_2021]. The
 list then gets cross-linked to the project's corresponding source control
 platforms (so that, we can reference the revision that we are fetching). Later
 on, the data gets downloaded via `git` and then gets sanitized until there is
@@ -164,8 +164,8 @@ for $target in ...:
 # Evaluation
 
 For the limited subset of things that can be described in any of the competitor engines, we
-evaluate the performance of `Reiz` by running similiar queries in Github Code Search \[@githubcodesearch\]
-grep.app \[@grepapp\] and Krugle \[@krugle\] and report back the amount of true / false positives. Each
[…]
