+++ paper/paper.bib
@@ -37,7 +37,7 @@
 @article{hugovankemenade_richardsi_2021,
     author    = {van Kemenade, Hugo and Si, Richard},
     publisher = {Zenodo},
-    date      = {2021-04},
+    date      = {2021},
     doi       = {10.5281/zenodo.4657163},
     title     = {hugovk/top-pypi-packages: Release 2021.04},
 }
@@ -61,7 +61,7 @@
 
 @misc{githubcodesearch,
     publisher = {GitHub},
-    url       = {https://grep.app},
+    url       = {https://github.com/search},
     date      = {2021},
     title     = {Github Code Search},
 }
@@ -89,7 +89,7 @@
 }
 
 @misc{fast,
-    publisher    = {jonatas},
+    author       = {Paganini, Jônatas Davi},
     url          = {https://github.com/jonatas/fast},
     date         = {2017},
     journaltitle = {GitHub repository},
@@ -97,7 +97,7 @@
 }
 
 @misc{astsearch,
-    publisher    = {takluyver},
+    author       = {Kluyver, Thomas},
     url          = {https://github.com/takluyver/astsearch},
     date         = {2014},
     journaltitle = {GitHub repository},
