+++ paper/generate_pdf.sh
@@ -1,0 +1,9 @@
+# build
+docker run --rm \
+    --volume $PWD:/data \
+    --user $(id -u):$(id -g) \
+    --env JOURNAL=joss \
+    openjournals/paperdraft
+
+# show
+firefox paper.pdf
