+++ paper/source/header.md
@@ -1,17 +1,0 @@
----
-title: 'Reiz: Structural Source Code Search'
-tags:
-    - Python
-    - source code search
-    - big code
-    - structure analysis
-authors:
-    - name: Batuhan Taskaya
-      affiliation: 1
-affiliations:
-    - name: Python Core Developer
-      index: 1
-
-date: 21 March 2021
-bibliography: paper.bib
----

[…]
